refactor to prepare for yml
