+++ backend/src/main/kotlin/no/bekk/util/RecordIDMapper.kt
@@ -3,13 +3,23 @@ package no.bekk.util
 import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.jsonPrimitive
 import no.bekk.configuration.getDatabaseConnection
-import no.bekk.services.AirTableService
+import no.bekk.providers.AirTableProvider
+import no.bekk.services.TableService
 import java.sql.Connection
 import java.sql.SQLException
 
 class RecordIDMapper {
 
-    suspend fun getIdMapFromAirTable(airTableService: AirTableService): Map<String, Pair<String, String>> {
+    suspend fun run() {
+        val tableService = TableService()
+        tableService.getTableProviders().forEach {
+            when (it) {
+                is AirTableProvider -> updateRecordIdsInDatabase(getIdMapFromAirTable(it))
+            }
+        }
+    }
+
+    private suspend fun getIdMapFromAirTable(airTableService: AirTableProvider): Map<String, Pair<String, String>> {
         val airTableData = airTableService.fetchDataFromMetodeverk()
         return airTableData.records.mapNotNull { record ->
             val questionId = record.fields.jsonObject["ID"]?.jsonPrimitive?.content
